moved all the functions for the virtual mesh into a file for creating a Class called VirtualMesh out of it later
