+++ src/simulation/utilities/error_checking.cpp
@@ -7,7 +7,6 @@
 #include <vector>
 #include <Eigen/Dense>
 
-#include <utilities/check_boundary.h>
 #include <utilities/check_validity.h>
 #include <utilities/sim_structs.h>
 
@@ -28,6 +27,25 @@ void error_invalid_values(
     if (checkForInvalidValues(r_UV_new)) {
         std::exit(1);  // stop script execution
     }
+}
+
+bool is_inside_uv(const Eigen::Vector2d& r) {
+    return (0 <= r[0] && r[0] <= 1) && (0 <= r[1] && r[1] <= 1);
+}
+
+std::vector<int> find_inside_uv_vertices_id(const Eigen::Matrix<double, Eigen::Dynamic, 2>& r) {
+    int nrows = r.rows();
+    std::vector<int> inside_id;
+
+    for (int i = 0; i < nrows; ++i) {
+        // Check if the point is inside the UV parametrization bounds
+        Eigen::Vector2d first_two_columns = r.row(i).head<2>();
+        if (is_inside_uv(first_two_columns)) {
+            inside_id.push_back(i);
+        }
+    }
+
+    return inside_id;
 }
 
 void error_lost_particles(
